@@ -1,10 +1,10 @@
 module github.com/rodaine/wrk
 
-go 1.23.9
+go 1.24.0
 
 require (
 	github.com/stretchr/testify v1.11.1
-	golang.org/x/sync v0.16.0
+	golang.org/x/sync v0.17.0
 )
 
 require (
